Exotic Audio Handling
Changes to how exotic audio is handled. Focused on True HD and Opus for right now (need samples of other formats), this changes the default behavior for Open, Import, and Add to Queue.
With a new Preference to allow disabling of exotic multi-channel tracks by default, those tracks will be deselected by default. If that option is not used, the tracks will now default to AAC + AC3, for more consistent handling. Stereo or Mono tracks are handled as just stereo tracks, no matter what.
This also addresses the scenario where an MP4 container has an AAC 6ch track, with no 2ch fallback. Subler wouldn't give you control over how this is handled when using Open or Add to Queue. This is now consistent across all paths.